Created Base classes instead of relying on decorator and class with static/class functions
