+++ kydb/__init__.py
@@ -1,12 +1,11 @@
 from .api import connect
-from .dbobj import dbobj, stored, DbObjManager
-from .exceptions import KydbException, DbObjException
+from .objdb import ObjDBMixin
+from .dbobj import DbObj, stored
 
 __all__ = [
     'connect',
-    'dbobj',
     'stored',
-    'DbObjManager',
-    'KydbException',
-    'DbObjException'
+    'ObjDBMixin',
+    'DbObj',
+    'stored'
 ]
