--- kydb/dbobj/__init__.py
@@ -1,3 +1,0 @@
-from .api import dbobj, stored, DbObjManager
-
-__all__ = ['dbobj', 'stored', 'DbObjManager']
